extract common code to its own mod
